requirements.txt
Django==2.1.3
djangorestframework==3.9.2
packaging==17.1
pylint==1.8.4
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.7
uWSGI==2.0.15

